Practice2 - Postgres - Image Data/requirements.txt
flask
flask-sqlalchemy
flask-smorest
flask_cors
python-dotenv
marshmallow